requirements.txt
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1603108395799/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028898966/work
brotlipy==0.7.0
certifi==2022.12.7
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1666798198223/work
cryptography @ file:///tmp/build/80754af9/cryptography_1652101588893/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy @ file:///tmp/build/80754af9/debugpy_1637091799509/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
docker-pycreds==0.4.0
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926439650/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fonttools==4.25.0
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1669279893622/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1676637751542/work
h5py @ file:///tmp/build/80754af9/h5py_1637138488546/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1647000773790/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///opt/conda/conda-bld/ipython_1651600145335/work
jedi @ file:///tmp/build/80754af9/jedi_1644297102865/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1633637554808/work
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1650622202839/work
jupyter-core @ file:///opt/conda/conda-bld/jupyter_core_1651671229925/work
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
kornia @ file:///home/conda/feedstock_root/build_artifacts/kornia_1652801734335/work
libtmux==0.21.0
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1647441664166/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
munkres==1.1.4
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847906199/work
numexpr @ file:///opt/conda/conda-bld/numexpr_1656071944761/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.20.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-model-summary @ file:///home/conda/feedstock_root/build_artifacts/pytorch-model-summary_1598394892458/work
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1648757097602/work
pyzmq @ file:///tmp/build/80754af9/pyzmq_1638434985866/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1652976687386/work
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1653073865807/work
sentry-sdk @ file:///home/conda/feedstock_root/build_artifacts/sentry-sdk_1675784087808/work
setproctitle @ file:///home/conda/feedstock_root/build_artifacts/setproctitle_1649637314806/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1611376390914/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1643647933166/work
torch==1.11.0
torch-summary==1.4.5
torchaudio==0.11.0
torchvision==0.12.0
tornado @ file:///tmp/build/80754af9/tornado_1606942317143/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1650891076910/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
wandb @ file:///home/conda/feedstock_root/build_artifacts/wandb_1675808907428/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
